SERVER-Spyhole/requirements.txt
flask
face_recognition
opnecv-python
pillow
numpy
